+++ app/views/planning_applications/review/review_immunity_enforcements/_shared_information.html.erb
