+++ app/views/planning_applications/review/review_permitted_development_rights/_shared_information.html.erb
