+++ config/routes.rb
@@ -177,19 +177,21 @@ Rails.application.routes.draw do
 
       resources :review_tasks, only: :index
 
-      resources :review_documents, only: %i[index] do
-        patch :update, on: :collection
+      scope module: :review do
+        resources :review_documents, only: %i[index] do
+          patch :update, on: :collection
+        end
+
+        resources :review_immunity_details, only: %i[edit update show]
+
+        resources :review_immunity_enforcements, only: %i[show edit update]
+
+        resources :review_local_policies, only: %i[edit update show]
+
+        resources :review_permitted_development_rights, only: %i[show edit update]
+
+        resources :review_policy_classes, only: %i[edit update show]
       end
-
-      resources :review_immunity_details, only: %i[edit update show]
-
-      resources :review_immunity_enforcements, only: %i[show edit update]
-
-      resources :review_local_policies, only: %i[edit update show]
-
-      resources :review_permitted_development_rights, only: %i[show edit update]
-
-      resources :review_policy_classes, only: %i[edit update show]
     end
   end
